Use HackerNews story title for timeline node titles
Changed HackerNews comment entries to display the story/thread title
instead of the comment text as the title. This makes it clearer what
discussion the comment is part of.

Before: 'I had not thought of it this way, but interesting point. ...'
After: 'AWS to bare metal two years later: Answering your questions...'

The comment text is still available in the summary and full content
when opening the detail panel.